==> Part 1 - Data Download.ipynb <==
# Project 3 - Part 1 - Data Download

- Name: Tyler Schelling

- Date Started: 12/13/2022

---

For this project, you have been hired to produce a MySQL database on Movies from a subset of IMDB's publicly available dataset. Ultimately, you will use this database to analyze what makes a movie successful and will provide recommendations to the stakeholder on how to make a successful movie. 

---

## Specifications

Your stakeholder only wants you to include information for movies based on the following specifications:

- Exclude any movie with missing values for genre or runtime
- Include only full-length movies (titleType = "movie").
- Include only fictional movies (not from documentary genre)
- Include only movies that were released 2000 - 2021 (include 2000 and 2021)
- Include only movies that were released in the United States

---

## Import Libraries

import os
import pandas as pd
import numpy as np

## Downloading the Files

basics_url = 'https://datasets.imdbws.com/title.basics.tsv.gz'
ratings_url = 'https://datasets.imdbws.com/title.ratings.tsv.gz'
akas_url = 'https://datasets.imdbws.com/title.akas.tsv.gz'

basics = pd.read_csv(basics_url, sep='\t', low_memory=False)
ratings = pd.read_csv(ratings_url, sep='\t', low_memory=False)
akas = pd.read_csv(akas_url, sep='\t', low_memory=False)

basics.head()

ratings.head()

akas.head()

## Filtering/Cleaning

#Replace null values ("\N") with np.nan across all 3 tables
basics = basics.replace({'\\N':np.nan})
ratings = ratings.replace({'\\N':np.nan})
akas = akas.replace({'\\N':np.nan})

#Eliminate movies that are null for runtimeMinutes
basics = basics[basics['runtimeMinutes'].notna()]

#Keep only titleType of 'Movie'
basics = basics[basics['titleType'] == 'movie']

#Eliminate movies that are null or contain documentary as their genre.
basics = basics[(basics['genres'].notna())]
basics = basics[~basics['genres'].str.contains('documentary',case=False)]

#Keep startYear 2000-2022
basics = basics[basics['startYear'].notna()]
basics['startYear'] = basics['startYear'].astype('int64')

year_filter = (basics['startYear'] >= 2000) & (basics['startYear'] < 2023)
basics = basics[year_filter]

#Keep only movies made in the US
akas = akas[akas['region'] == 'US']
us_filter_basics = basics['tconst'].isin(akas['titleId'])
us_filter_ratings = ratings['tconst'].isin(akas['titleId'])

#Filter the basics DF for US movies
basics = basics[us_filter_basics]
#Filter the ratings DF for US movies
ratings = ratings[us_filter_ratings]

## Saving the Filtered Tables to the Repository

#Create Data folder in Repository to store CSV files
os.makedirs('Data/',exist_ok=True) 
# Confirm folder created
os.listdir("Data/")

## Save current dataframe to file.
basics.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)
ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)
akas.to_csv("Data/title_akas.csv.gz",compression='gzip',index=False)

## Open Filtered Tables 

# Open saved file and preview again
basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)
basics.head()

# Open saved file and preview again
ratings = pd.read_csv("Data/title_ratings.csv.gz", low_memory = False)
ratings.head()

# Open saved file and preview again
akas = pd.read_csv("Data/title_akas.csv.gz", low_memory = False)
akas.head()

## Display Info of each Dataframe

basics.info()

ratings.info()

akas.info()
