==> Part 3 - Database ETL.ipynb <==
# Project 3 - Part 3 - Database ETL

- Name: Tyler Schelling
- Date Started: 1/17/2023

---


## Specifications - Database

- Your stakeholder wants you to take the data you have been cleaning and collecting in Parts 1 & 2 of the project, and wants you to create a MySQL database for them.

- Specifically, they want the data from the following files included in your database:
    - Title Basics:
        - Movie ID (tconst)
        - Primary Title
        - Start Year
        - Runtime (in Minutes)
        - Genres
    - Title Ratings
        - Movie ID (tconst)
        - Average Movie Rating
        - Number of Votes
    - The TMDB API Results (multiple files)
        - Movie ID
        - Revenue
        - Budget
        - Certification (MPAA Rating)
- You should normalize the tables as best you can before adding them to your new database.
    - Note: an important exception to their request is that they would like you to keep all of the data from the TMDB API in 1 table together (even though it will not be perfectly normalized).
    - You only need to keep the imdb_id, revenue, budget, and certification columns

## Import Libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import json

import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy.types import *
from sqlalchemy_utils import create_database, database_exists
from sqlalchemy import create_engine

## Normalize Genres

### title_genres Dataframe

#Load the dataframe from the csv file
title_basics_df = pd.read_csv('Data/title_basics.csv.gz')

#Display the first 5 rows
title_basics_df.head()

#Save tconst and genres to new df
title_genres_df = title_basics_df[['tconst']].copy()
title_genres_df['genres split'] = title_basics_df['genres'].str.split(',')

#Display the first 5 rows
title_genres_df.head()

#Update the df to have unique lines for each genre
title_genres_df = title_genres_df.explode('genres split')

#Display the first 5 rows
title_genres_df.head()

#Create a dictionary of unique genres of keys and apply an integer as the value
unique_genres = sorted(title_genres_df['genres split'].unique())
genre_map = dict(zip(unique_genres, range(len(unique_genres))))
genre_map

#Map the genre_id to the genre names and drop the genres split column
title_genres_df['genre_id'] = title_genres_df['genres split'].map(genre_map)
title_genres_df.drop(columns = 'genres split', inplace = True)

#Display the first 5 rows
title_genres_df.head()

### genres Dataframe

#Create a dataframe from the genre_map dictionary
genres_df = pd.DataFrame({'genre_name': genre_map.keys(), 
                          'genre_id': genre_map.values()})

#Display the first 5 rows
genres_df.head()

## Update title_basics

#Display dataframe info
title_basics_df.info()

#Drop columns from the dataframe
#"titleType": every row will be a movie
#"original_title": we will use the primary title column instead
#"isAdult": "Adult" will show up in the genres so this is redundant information
#"endYear": all rows are NaN
#"genres": genre is now represented in the 2 new tables described in 3.3
title_basics_df = title_basics_df.drop(columns = ['titleType', 'originalTitle',
                                                  'isAdult', 'endYear', 'genres'])

#Display dataframe info
title_basics_df.info()

#Display the first 5 rows
title_basics_df.head()

## Create SQL Movies Database

#Obtain the MySQL login username and password
with open('/Users/tyler/.secret/mysql.json') as f:
    login = json.load(f)
#Display the MySQL login keys
login.keys()

#Create a connection string using credentials above
from urllib.parse import quote_plus as urlquote
database_name = "Movies"
connection = f"mysql+pymysql://{login['username']}:{login['password']}@localhost/{database_name}"

#Create an instance of sqlalchemy 
engine = create_engine(connection)

#Create the Movies database, if it does not already exist
if database_exists(connection):
    print('It exists!')
else:
    create_database(connection)
    print('Database created!')

### title_basics SQL Table

#Create a schema dictonary using datatype objects
df_schema = {
    'tconst': String(title_basics_df['tconst'].str.len().max()),
    'primaryTitle': Text(title_basics_df['primaryTitle'].str.len().max()),
    'startYear': Integer(),
    'runtimeMinutes': Integer()}

#Save the dataframe to an sql table
title_basics_df.to_sql('title_basics',
              engine, 
              dtype=df_schema,
              if_exists='replace',
              index=False)

#Use the engine to update title_basics with a primary key
engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')

#Display the first 5 rows of the table using an SQL query.
q = """
SELECT * 
FROM title_basics LIMIT 5
;"""

pd.read_sql(q, engine)

### title_ratings SQL Table

#Load the dataframe from the csv file
title_ratings_df = pd.read_csv('Data/title_ratings.csv.gz')

#Display the first 5 rows
title_ratings_df.head()

#Create a schema dictonary using datatype objects
df_schema = {
    "tconst": String(title_ratings_df['tconst'].str.len().max()), 
    'averageRating':Float(),
    'numVotes':Integer()}

#Save the dataframe to an sql table
title_ratings_df.to_sql('title_ratings',
              engine, 
              dtype=df_schema,
              if_exists='replace',
              index=False)

#Use the engine to update title_basics with a primary key
engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')

#Display the first 5 rows of the table using a SQL query
q = """
SELECT * 
FROM title_ratings LIMIT 5
;"""

pd.read_sql(q, engine)

### title_genres SQL Table

#Display the first 5 rows
title_genres_df.head()

#Create a schema dictonary using datatype objects
df_schema = {
    "tconst": String(title_genres_df['tconst'].str.len().max()), 
    'genre_id':Integer()}

#Save the dataframe to an sql table
title_genres_df.to_sql('title_genres',
              engine, 
              dtype=df_schema,
              if_exists='replace',
              index=False)

#Display the first 5 rows of the table using a SQL query
q = """
SELECT * 
FROM title_genres LIMIT 5
;"""

pd.read_sql(q, engine)

### genres SQL Table

#Display the first 5 rows
genres_df.head()

#Create a schema dictonary using datatype objects
df_schema = {
    "genre_id": Integer(), 
    'genre_name': String(genres_df['genre_name'].str.len().max())}

#Save the dataframe to an sql table
genres_df.to_sql('genres',
              engine, 
              dtype=df_schema,
              if_exists='replace',
              index=False)

#Use the engine to update title_basics with a primary key
engine.execute('ALTER TABLE genres ADD PRIMARY KEY (`genre_id`);')

#Display the first 5 rows of the table using an SQL query
q = """
SELECT * 
FROM genres LIMIT 5
;"""

pd.read_sql(q, engine)

### tmdb_data SQL Table

#Load the dataframe from the csv file
tmdb_data_df = pd.read_csv('Data/tmdb_results_combined.csv.gz')

#Display the first 5 rows
tmdb_data_df.head()

#Drop irrelevant columns
tmdb_data_df = tmdb_data_df.loc[:,["imdb_id","revenue",'budget','certification']]
tmdb_data_df.drop(0, inplace=True, axis=0)

#Display the first 5 rows
tmdb_data_df.head()

#Create a schema dictonary using datatype objects
df_schema = {
    'imdb_id': String(tmdb_data_df['imdb_id'].str.len().max()), 
    'budget':Float(),
    'revenue':Float(),
    'certification': String(tmdb_data_df['certification'].str.len().max())}

#Save the dataframe to an sql table
tmdb_data_df.to_sql('tmdb_data',
              engine, 
              dtype=df_schema,
              if_exists='replace',
              index=False)

#Use the engine to update title_basics with a primary key
engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

#Display the first 5 rows of the table using an SQL query
q = """
SELECT * 
FROM tmdb_data LIMIT 5
;"""

pd.read_sql(q, engine)

## Show SQL Tables

#Display the SQL table names using an SQL query
q = """SHOW TABLES;"""
pd.read_sql(q, engine)
